common js structure created
